fix(scheduler): widen MT4_TRADES.TIMESTAMP to i64
Deployed and immediately hit "out of range integral type conversion"
decoding TIMESTAMP as i32 for real service-20 rows -- real values exceed
i32::MAX. MT4 TIMESTAMP is an ever-growing internal modification counter
on busy servers, not a plain unix-seconds value as the int(11) column
type suggested.

+++ services/scheduler/src/db/mt4.rs
@@ -122,8 +122,11 @@ pub struct Mt4ClosedTrade {
     pub swaps: f64,
     #[sqlx(rename = "REASON")]
     pub reason: i32,
+    // MT4's TIMESTAMP is documented as int(11) but is effectively an ever-growing
+    // internal modification counter on busy servers, not a plain unix-seconds
+    // value — real values have been observed to exceed i32::MAX. Widen to i64.
     #[sqlx(rename = "TIMESTAMP")]
-    pub timestamp: i32,
+    pub timestamp: i64,
 }
 
 /// 查询新的已平仓 trades（cursor-based 分页，供 TradeMonitor 轮询）。
